OPT: Move Downloader & Version to Resource namespace as it's not backed by DB
